+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/proxies/index.ts
@@ -3,5 +3,5 @@
  * Licensed under the MIT License.
  */
 
-export { getProxyForField } from "./node";
+export { getProxyForField } from "./proxies";
 export { List } from "./types";
